requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
openai==1.3.5
PyPDF2==3.0.1
python-docx==0.8.11
SQLAlchemy==2.0.23
